Change Database Design
The nested serializers wouldnt allow me to update the available times
under the previous design. I will have to revert to that design, but for
now im just using a comma delimited string, which I will parse in each
application. To update, I will POST a new string to the field, which
does not include the time that was booked.